smoothly converting axis events to touch events

+++ frameworks/core/pipeline_ng/pipeline_context.cpp
@@ -3442,7 +3442,7 @@ void PipelineContext::OnTouchEvent(const TouchEvent& point, const RefPtr<FrameNo
     ACE_BENCH_MARK_TRACE("OnTouchEvent_start type:%d", static_cast<int32_t>(point.type));
 
     TAG_LOGD(AceLogTag::ACE_INPUTKEYFLOW, "OnTouchEvent type:%{public}lu, isGenerate:%{public}d",
-        point.type, point.isGenerate);
+        static_cast<int32_t>(point.type), point.isGenerate);
     if (ConvertFromMouseAxis(point) && !point.isGenerate && compatibleManager_.NotifyNewEvent(point)) {
         return;
     }
